UCP/WIREUP: propogated priority into uct_ep_params in all wireups

+++ src/ucp/core/ucp_types.h
@@ -50,7 +50,7 @@ UCP_UINT_TYPE(UCP_MAX_LANES)         ucp_lane_map_t;
 
 /* Message Priority */
 #define UCP_MAX_PRIORITIES 2     /* Currently only Low (0) and High (1) are available */
-typedef uint16_t                     ucp_priority_t;
+typedef uint8_t ucp_priority_t;
 
 /* System devices */
 #define UCP_MAX_SYS_DEVICES          64
